Fix imports and add missing eslint package
