+++ package.json
@@ -35,6 +35,7 @@
     "babel-plugin-root-import": "6.6.0",
     "eslint": "^7.17.0",
     "eslint-config-prettier": "^7.1.0",
+    "eslint-plugin-jsx-a11y": "^6.4.1",
     "gatsby-plugin-ts-config": "^1.1.5",
     "prettier": "2.2.1",
     "typescript": "4.1.3"
