+++ src/pages/login.tsx
@@ -1,10 +1,10 @@
 import React from 'react'
 import {PageProps} from 'gatsby'
-import LoginSection from "@/components/LoginSection";
+import LoginForm from "@/components/LoginForm";
 
 const Login: React.FC<PageProps> = () => (
     <main>
-        <LoginSection />
+        <LoginForm />
     </main>
 )
